Initial Commit
This includes more accurate instructions regarding making install disks
for Macs. It includes instructions on how to burn DVDs as well because
older Macs are incapable of booting elementary from USB. The only issue
is that the long Mac documentation causes the Windows and Linux tabs to
have lots of whitespace. I have no idea how to fix that.