Getting Started guide example fixes (#1587)
* Getting Started guide example fixes

The "Functionality in Gtk.Grid" section introduces the `this` concept at the end of the example code:

```vala
this.add (grid);
```

but it's the first time it's mentioned, the guide uses ` window.add (grid)` in all other examples and it does not explain where the `this` comes from.

If I understand correctly the section is supposed to explain communication between the button and the label. Now the code to update said label:

```vala
button.clicked.connect (() => {
  button.label = _("Hello World!");
  button.sensitive = false;
});
```

This will update the text on the button, not the label created earlier. It should be:

```vala
button.clicked.connect (() => {
  label.set_text (_("Hello World!"));
  button.sensitive = false;
});
```

* Use property instead of setter