Optimize images (#1588)
* update package versions

* update stylelint rules

* run all screenshots by optipng

* animate toast exit (#1578)

* animate toast exit

* Update main.js

* Make toast responsive for small mobile displays (#1580)

* Make toast responsive for small mobile displays

* fix bug in iOS Safari

* try try again

* iOS pls

* cmon bby dont do me lyk dat

* lint

* fix javascript issues (#1581)

* fix dxtra close tag (#1582)

* replace preview image and add svg source (#1586)

* Getting Started guide example fixes (#1587)

* Getting Started guide example fixes

The "Functionality in Gtk.Grid" section introduces the `this` concept at the end of the example code:

```vala
this.add (grid);
```

but it's the first time it's mentioned, the guide uses ` window.add (grid)` in all other examples and it does not explain where the `this` comes from.

If I understand correctly the section is supposed to explain communication between the button and the label. Now the code to update said label:

```vala
button.clicked.connect (() => {
  button.label = _("Hello World!");
  button.sensitive = false;
});
```

This will update the text on the button, not the label created earlier. It should be:

```vala
button.clicked.connect (() => {
  label.set_text (_("Hello World!"));
  button.sensitive = false;
});
```

* Use property instead of setter

* fix button offset when already paid

* move notebook image to jpg

* use jpg for all screenshots

* inline press svg images

* run all svg files with svgo

* update svg files with svgo

* move images to _images folder

* clean travis file

* build all image files

* move building styles to gulp task instead of webpack

* remove watch scripts

* update webpack to include source maps

* add js used icons to gulp build list

* polyfill a bit better for scripts