More coherent licenses in Getting Started doc
The Copying paragraph talks about the GPL and instructs the developer to
create a COPYING file mentioning the GPL.

```
For elementary apps this is typically the GNU General Public License (GPL).
```

The newer AppData.xml paragrah shows how to specify a license in this
file but uses the CC0 license as an example:

```xml
<metadata_license>CC0</metadata_license>
```

Wouldn't it make more sense to use the same license in both places? I
feel we should not send people down the CC0 versus GPL rabbit hole when
they are simply trying to create their first Vala app.