Make focus indicators consistent
Focus indicators are important for keyboard users, to navigate around the website, without them it is very difficult or impossible to know where you are on the page.

This change removes the styles that prevent focus indicators being shown and bring them inline with the rest of the website e.g. blog page.